refactor(components): checkbox-wrapper and radio-button utils | tg | #2487
