+++ packages/components/src/components/checkbox-wrapper/checkbox-wrapper.spec.ts
@@ -1,5 +1,5 @@
 import { CheckboxWrapper } from './checkbox-wrapper';
-import * as checkboxButtonWrapperUtils from '../../utils/checkbox-radio-button-wrapper/checkbox-radio-button-wrapper-utils';
+import * as checkboxButtonWrapperUtils from '../../utils/checkbox-radio-button-wrapper-utils';
 
 describe('componentWillLoad', () => {
   it('should call addChangeListener() with correct parameters', () => {
