+++ packages/components/src/components/checkbox-wrapper/checkbox-wrapper.tsx
@@ -19,7 +19,7 @@ import { getComponentCss } from './checkbox-wrapper-styles';
 import type { CheckboxWrapperState } from './checkbox-wrapper-utils';
 import { StateMessage } from '../common/state-message/state-message';
 import { Required } from '../common/required/required';
-import { addChangeListener } from '../../utils/checkbox-radio-button-wrapper/checkbox-radio-button-wrapper-utils';
+import { addChangeListener } from '../../utils/checkbox-radio-button-wrapper-utils';
 
 const propTypes: PropTypes<typeof CheckboxWrapper> = {
   label: AllowedTypes.string,
