+++ packages/components/src/components/radio-button-wrapper/radio-button-wrapper.tsx
@@ -19,7 +19,7 @@ import { getComponentCss } from './radio-button-wrapper-styles';
 import { StateMessage } from '../common/state-message/state-message';
 import { Required } from '../common/required/required';
 import type { RadioButtonWrapperState } from './radio-button-wrapper-utils';
-import { addChangeListener } from '../../utils/checkbox-radio-button-wrapper/checkbox-radio-button-wrapper-utils';
+import { addChangeListener } from '../../utils/checkbox-radio-button-wrapper-utils';
 
 const propTypes: PropTypes<typeof RadioButtonWrapper> = {
   label: AllowedTypes.string,
